ariaConfig object processing moved to own file
